requirements.txt
# Python Version: 3.9.11

# Package Versions
tensorflow==2.13.0
keras==2.13.1
streamlit==1.33.0
Pillow==10.3.0
numpy==1.24.3